+++ evals/characteristics/src/main.py
@@ -4,220 +4,11 @@ import json
 import logging
 from pathlib import Path
 
-import yaml
-from jinja2 import Environment, FileSystemLoader, select_autoescape
-from pydantic import BaseModel
-
-from common.llm.client import ChatBot, create_chatbot
-from evals.characteristics.src.types import (
-    CharacteristicDetection,
-    CharacteristicExtractionOutput,
-    EvalsConfig,
-    ExtractionMetadata,
-    ProcessedFileResult,
-)
-
-"""
-Characteristic Extraction evaluation pipeline.
-
-This script processes transcripts (TXT or JSON) to extract specific characteristics
-using gpt4o on APIM. It handles large transcripts by chunking them with an overlap,
-sanitizes input to avoid WAF issues, and maps detected characteristics back
-to their original positions in the transcript.
-"""
+from evals.characteristics.src.config_loader import load_config
+from evals.characteristics.src.pipeline import process_file
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
-
-
-class TranscriptTurn(BaseModel):
-    """Represents a single turn in a conversation transcript."""
-
-    speaker: str
-    text: str
-
-
-def load_config(config_path: Path) -> EvalsConfig:
-    """
-    Loads the evaluation configuration from a YAML file.
-
-    Args:
-        config_path: Path to the YAML configuration file.
-
-    Returns:
-        An EvalsConfig object containing the evaluation settings.
-    """
-    if not config_path.exists():
-        logger.warning("Config file %s not found. Using default parameters.", config_path)
-        return EvalsConfig()
-
-    with config_path.open("r", encoding="utf-8") as f:
-        raw_yaml = yaml.safe_load(f) or {}
-    return EvalsConfig(**raw_yaml)
-
-
-def render_prompt(template_path: str, transcript: str) -> str:
-    """
-    Renders a Jinja2 template with the provided transcript.
-
-    Args:
-        template_path: Path to the Jinja2 template file.
-        transcript: The transcript text to include in the prompt.
-
-    Returns:
-        The rendered prompt string.
-    """
-    path = Path(template_path).resolve()
-    env = Environment(
-        loader=FileSystemLoader(path.parent),
-        autoescape=select_autoescape(),
-    )
-    template = env.get_template(path.name)
-    return template.render(transcript=transcript)
-
-
-def load_transcript(file_path: Path) -> str:
-    """
-    Loads and formats transcript from a file.
-    Supports .txt (raw text) and .json (list of speaker/text pairs).
-
-    Args:
-        file_path: Path to the transcript file.
-
-    Returns:
-        A single string representing the full transcript.
-    """
-    # 1. Handle raw text as a baseline
-    if file_path.suffix == ".txt":
-        return file_path.read_text(encoding="utf-8")
-
-    # 2. Assume compliant JSON (list of {speaker: ..., text: ...})
-    if file_path.suffix == ".json":
-        raw_data = json.loads(file_path.read_text(encoding="utf-8"))
-        # Using Pydantic to validate the input structure immediately
-        turns = [TranscriptTurn(**item) for item in raw_data]
-        return "\n".join(f"{t.speaker}: {t.text}" for t in turns)
-
-    message = f"Unsupported format: {file_path.suffix}"
-    logger.error(message)
-    raise ValueError(message)
-
-
-_WAF_TRANSLATION_TABLE = str.maketrans("-\u2013\u2014", "   ", "<>*;={}()`~|")
-
-
-def sanitize_for_waf(text: str) -> str:
-    """
-    Aggressive sanitizer to bypass Azure WAF OWASP False Positives.
-    Removes characters that trigger SQLi and XSS anomaly rules.
-    """
-    return text.translate(_WAF_TRANSLATION_TABLE)
-
-
-def deduplicate_characteristics(characteristics: list[CharacteristicDetection]) -> list[CharacteristicDetection]:
-    """
-    Removes duplicate characteristics based on their category and value.
-    This is necessary because overlapping chunks may result in duplicate extractions.
-
-    Args:
-        characteristics: A list of characteristic objects.
-
-    Returns:
-        A list of unique characteristic objects.
-    """
-    unique_characteristics = []
-    seen_signatures = set()
-
-    for item in characteristics:
-        category = item.characteristic
-        value = item.attribute_value
-        signature = f"{category}|{value}"
-
-        if signature not in seen_signatures:
-            seen_signatures.add(signature)
-            unique_characteristics.append(item)
-
-    return unique_characteristics
-
-
-def build_chunks(transcript: str, chunk_size_chars: int = 1000, overlap_chars: int = 250) -> list[tuple[str, int]]:
-    stride = chunk_size_chars - overlap_chars
-    chunks = []
-    start = 0
-    while start < len(transcript):
-        chunks.append((transcript[start : start + chunk_size_chars], start))
-        start += stride
-    return chunks
-
-
-async def process_chunk(
-    chunk_text: str, offset: int, prompt_path: Path, chatbot: ChatBot
-) -> list[CharacteristicDetection]:
-    safe_chunk = sanitize_for_waf(chunk_text)
-    prompt_text = render_prompt(str(prompt_path), safe_chunk)
-    response = await chatbot.structured_chat([{"role": "user", "content": prompt_text}], CharacteristicExtractionOutput)
-    for item in response.detected_characteristics:
-        for span in item.evidence_spans:
-            if span.start_index is not None:
-                span.start_index += offset
-            if span.end_index is not None:
-                span.end_index += offset
-    return response.detected_characteristics
-
-
-async def process_file(file_path: Path, config: EvalsConfig, root_dir: Path) -> ProcessedFileResult:
-    """
-    Processes a single transcript file, extracting characteristics in chunks.
-
-    Args:
-        file_path: Path to the input transcript.
-        config: Evaluation configuration.
-        root_dir: The project root directory.
-
-    Returns:
-        A ProcessedFileResult containing the extraction results and metadata.
-    """
-    prompt_rel_path = Path(config.prompts.extraction_template)
-    template_name = prompt_rel_path.stem
-    prompt_path = root_dir / prompt_rel_path
-    model_name = config.model.model
-
-    logger.info("Processing %s using model '%s'", file_path.name, model_name)
-
-    transcript = load_transcript(file_path)
-    chatbot = create_chatbot(
-        model_type=config.model.provider,
-        model_name=config.model.model,
-        temperature=config.model.temperature,
-    )
-
-    all_detected_characteristics: list[CharacteristicDetection] = []
-    failed_chunks: list[int] = []
-    chunks = build_chunks(transcript)
-
-    for idx, (chunk_text, offset) in enumerate(chunks):
-        logger.info("  -> Sending Chunk %d of %d to Azure...", idx + 1, len(chunks))
-        try:
-            detected = await process_chunk(chunk_text, offset, prompt_path, chatbot)
-            all_detected_characteristics.extend(detected)
-            logger.info("     Found %d characteristics in chunk %d", len(detected), idx + 1)
-        except (ValueError, RuntimeError, ConnectionError, TimeoutError) as e:
-            failed_chunks.append(idx + 1)
-            logger.error("     Failed on chunk %d: %s", idx + 1, e)
-        if idx < len(chunks) - 1:
-            await asyncio.sleep(2)
-
-    metadata = ExtractionMetadata(
-        model_used=model_name,
-        prompt_version=template_name,
-        total_chunks_processed=len(chunks),
-        failed_chunks=failed_chunks,
-    )
-    return ProcessedFileResult(
-        version="1.0",
-        detected_characteristics=deduplicate_characteristics(all_detected_characteristics),
-        metadata=metadata,
-    )
 
 
 async def main() -> None:
@@ -230,19 +21,12 @@ async def main() -> None:
         help="Path to the YAML configuration file",
     )
     args = parser.parse_args()
-
-    # Determine project root to resolve relative paths in config
-    root_dir = Path(__file__).resolve().parent.parent.parent.parent
-    if not (root_dir / "pyproject.toml").exists():
-        root_dir = Path.cwd()
-
-    # Load configuration
+    root_dir = Path(__file__).resolve().parents[3]
     config_path = Path(args.config)
     if not config_path.is_absolute():
         config_path = root_dir / config_path
     config = load_config(config_path)
 
-    # Set up directories
     input_dir = root_dir / config.dataset.input_dir
     output_dir = root_dir / config.run.output_dir
     output_dir.mkdir(parents=True, exist_ok=True)
@@ -254,13 +38,11 @@ async def main() -> None:
         )
         return
 
-    # Process all supported files in the input directory
     for file_path in input_dir.iterdir():
         if file_path.is_file() and file_path.suffix in [".txt", ".json"]:
             try:
                 result = await process_file(file_path, config, root_dir)
 
-                # Save results to output directory
                 output_file = output_dir / f"{file_path.stem}_output.json"
                 output_file.write_text(json.dumps(result.model_dump(), indent=2))
                 logger.info("Saved result to %s", output_file)
